% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/installSnapkin.R
\name{installSnapkin}
\alias{installSnapkin}
\title{SnapKin Conda Dependencies Installer}
\usage{
installSnapkin(useGPU = FALSE)
}
\arguments{
\item{useGPU}{Boolean specifying whether a GPU-supported environment should
be downloaded. This is not supported for MacOS environments.}
}
\description{
This function installs a Conda environment for SnapKin.
}
\examples{

installSnapkin()

}
